requirements.txt
flask==2.3.3
opencv-contrib-python==4.7.0.72
numpy==1.21.6
gunicorn==20.1.0
Pillow==9.5.0
